cleanup

+++ WWCrossFeed/WWCrossFeed/MainWindow.xaml.cs
@@ -18,7 +18,7 @@ using System.Windows.Shapes;
 
 namespace WWCrossFeed {
     public partial class MainWindow : Window {
-        private const float CAMERA_DISTANCE_DEFAULT = 10.0f;
+        private const double CAMERA_DISTANCE_DEFAULT = 10.0;
         private const double SMALLEST_ROOM_LENGTH = 2.0;
         private WWRoom mRoom = new WWRoom();
         private WWRoomVisualizer mRoomVisualizer;
@@ -91,7 +91,7 @@ namespace WWCrossFeed {
                 MessageBox.Show("Error: Camera parameter parse error");
                 return false;
             }
-            mRoomVisualizer.CameraFovHDegree = (float)fovH;
+            mRoomVisualizer.CameraFovHDegree = fovH;
             return true;
         }
